refactor: MVC 패턴 적용
- 패키지 구조 변경(domain, view 추가)
- Game의 run() 메서드 제거
  - 해당 메서드 역할은 main()이 대신 수행
- Game 테스트 코드 내 메서드 이름 변경
  - getWinPosition() > getMaxPosition()
