+++ src/main/java/Main.java
@@ -1,13 +1,26 @@
-import racingcar.Game;
-import racingcar.IOHelper;
+import racingcar.domain.Game;
+import racingcar.dto.GameInfo;
+import racingcar.view.IOHelper;
+
+import java.util.List;
 
 public class Main {
     public static void main(String[] args) {
         IOHelper ioHelper = new IOHelper();
 
         Game game = new Game();
-        game.run(ioHelper);
 
-        ioHelper.close();
+        List<String> namesInput = ioHelper.getNamesInput();
+        int roundInput = ioHelper.getRoundInput();
+
+        GameInfo gameInfo = game.init(namesInput, roundInput);
+
+        ioHelper.printInitialStatus(gameInfo);
+        while(gameInfo.getLeftRoundCnt() != 0) {
+            gameInfo = game.runRound();
+            ioHelper.printRoundResult(gameInfo);
+        }
+
+        ioHelper.printGameResult(game.findWinners(gameInfo));
     }
 }
