+++ src/main/java/racingcar/dto/CarInfo.java
@@ -1,4 +1,6 @@
-package racingcar;
+package racingcar.dto;
+
+import racingcar.domain.Car;
 
 import java.util.List;
 import java.util.stream.Collectors;
